move basic tests to core (and use async await)
